formateo info html movido hacia vista
