+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/model/repository/external/lastFM/LastFMToArtistResolver.kt
@@ -36,7 +36,7 @@ internal class JsonToArtistResolver : LastFMToArtistResolver {
         artistName = name
     }
     private fun initFormatArtistInfoHelper() {
-        formatArtistInfoHelper = LastFMInjector.getFormatArtistInfoHelper(artistName)
+        formatArtistInfoHelper = LastFMInjector.getFormatArtistInfoHelper()
     }
 
     private fun initQueryArtistInfo(serviceData: String?) {
